requirements.txt
pandas>=2.2
numpy>=1.26
pydantic>=2.6
python-dotenv>=1.0
pytest>=8.0
openpyxl>=3.1
pyarrow>=15.0
streamlit>=1.37
langchain>=0.2
openai>=1.40
faiss-cpu>=1.8
prophet>=1.1
pyngrok
plotly>=5.22
weasyprint>=59.0
jinja2
statsmodels>=0.14
langchain-community>=0.2
langchain-openai>=0.1
